LayoutTests/imported/w3c/web-platform-tests/css/css-break/hit-test-inline-fragmentation-with-border-radius-expected.txt

FOO
BAR
FOO
BAR
FOO
BAR

FAIL Horizontal line 1, hit test outside top-left rounded corner assert_equals: expected Element node <div>
<span id="horizontal" class="round">FOO<br>BAR</spa... but got Element node <span id="horizontal" class="round">FOO<br>BAR</span>
PASS Horizontal line 1, hit test inside top-left rounded corner
FAIL Horizontal line 1, hit test outside bottom-left rounded corner assert_equals: expected Element node <div>
<span id="horizontal" class="round">FOO<br>BAR</spa... but got Element node <span id="horizontal" class="round">FOO<br>BAR</span>
PASS Horizontal line 1, hit test inside bottom-left rounded corner
PASS Horizontal line 1, hit test inside top-right right-angled corner
PASS Horizontal line 1, hit test inside bottom-right right-angled corner
FAIL Horizontal line 2, hit test outside top-right rounded corner assert_equals: expected Element node <div>
<span id="horizontal" class="round">FOO<br>BAR</spa... but got Element node <span id="horizontal" class="round">FOO<br>BAR</span>
PASS Horizontal line 2, hit test inside top-right rounded corner
FAIL Horizontal line 2, hit test outside bottom-right rounded corner assert_equals: expected Element node <div>
<span id="horizontal" class="round">FOO<br>BAR</spa... but got Element node <span id="horizontal" class="round">FOO<br>BAR</span>
PASS Horizontal line 2, hit test inside bottom-right rounded corner
PASS Horizontal line 2, hit test inside top-left right-angled corner
PASS Horizontal line 2, hit test inside bottom-left right-angled corner
FAIL Vertical LR line 1, hit test outside top-left rounded corner assert_equals: expected Element node <div style="writing-mode:vertical-lr">
<span id="vertical... but got Element node <span id="vertical-lr" class="round">FOO<br>BAR</span>
PASS Vertical LR line 1, hit test inside top-left rounded corner
FAIL Vertical LR line 1, hit test outside top-right rounded corner assert_equals: expected Element node <div style="writing-mode:vertical-lr">
<span id="vertical... but got Element node <span id="vertical-lr" class="round">FOO<br>BAR</span>
PASS Vertical LR line 1, hit test inside top-right rounded corner
PASS Vertical LR line 1, hit test inside bottom-left right-angled corner
PASS Vertical LR line 1, hit test inside bottom-right right-angled corner
FAIL Vertical LR line 2, hit test outside bottom-left rounded corner assert_equals: expected Element node <div style="writing-mode:vertical-lr">
<span id="vertical... but got Element node <span id="vertical-lr" class="round">FOO<br>BAR</span>
PASS Vertical LR line 2, hit test inside bottom-left rounded corner
FAIL Vertical LR line 2, hit test outside bottom-right rounded corner assert_equals: expected Element node <div style="writing-mode:vertical-lr">
<span id="vertical... but got Element node <span id="vertical-lr" class="round">FOO<br>BAR</span>
PASS Vertical LR line 2, hit test inside bottom-right rounded corner
PASS Vertical LR line 2, hit test inside top-left right-angled corner
PASS Vertical LR line 2, hit test inside top-right right-angled corner
FAIL Vertical RL line 1, hit test outside top-left rounded corner assert_equals: expected Element node <div style="writing-mode:vertical-rl">
<span id="vertical... but got Element node <span id="vertical-rl" class="round">FOO<br>BAR</span>
PASS Vertical RL line 1, hit test inside top-left rounded corner
FAIL Vertical RL line 1, hit test outside top-right rounded corner assert_equals: expected Element node <div style="writing-mode:vertical-rl">
<span id="vertical... but got Element node <span id="vertical-rl" class="round">FOO<br>BAR</span>
PASS Vertical RL line 1, hit test inside top-right rounded corner
PASS Vertical RL line 1, hit test inside bottom-left right-angled corner
PASS Vertical RL line 1, hit test inside bottom-right right-angled corner
FAIL Vertical RL line 2, hit test outside bottom-left rounded corner assert_equals: expected Element node <div style="writing-mode:vertical-rl">
<span id="vertical... but got Element node <span id="vertical-rl" class="round">FOO<br>BAR</span>
PASS Vertical RL line 2, hit test inside bottom-left rounded corner
FAIL Vertical RL line 2, hit test outside bottom-right rounded corner assert_equals: expected Element node <div style="writing-mode:vertical-rl">
<span id="vertical... but got Element node <span id="vertical-rl" class="round">FOO<br>BAR</span>
PASS Vertical RL line 2, hit test inside bottom-right rounded corner
PASS Vertical RL line 2, hit test inside top-left right-angled corner
PASS Vertical RL line 2, hit test inside top-right right-angled corner
